MessageRestController modified and chat.js fixed
Modified MessageRestController to assure that the sender of a message is the user logged in and the getMessages method now return a list of messages instead of a page. Also, chat.js has been fixed so the messages that you write are shown correctly in the chat once you reenter.